Backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
django-filter>=24.2,<25
psycopg[binary]>=3.1,<4
gunicorn>=21.2,<22
